pyproject.toml
[tool.poetry]
name = "ros_pc2i"
version = "0.1.0"
description = ""
authors = ["HiraiKyo <kyo.mikhael@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
moderngl = "^5.12.0"
numpy = "^1.24"
numpy-stl = "^3.1"
pillow = "^10.4"
pyrr = "^0.10.3"
pyyaml = "^6.0.2"
glfw = "^2.8.0"
pyglm = "2.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
